refactor: 스타일 관련 코드 분리
